+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-v1.0.0-beta.42 - [December 3, 2022](https://github.com/lando/vuepress-theme-default-plus/releases/tag/v1.0.0-beta.42)
+v1.0.0-beta.42 - [December 4, 2022](https://github.com/lando/vuepress-theme-default-plus/releases/tag/v1.0.0-beta.42)
 -----------------------------------
 
 * Added `satisfies` to `plugin-versions-page` to target specific releases